+++ mirage-support/serializers/github-file.js
@@ -1,5 +1,4 @@
 import EmberDataGithubSerializer from './ember-data-github';
 
 export default EmberDataGithubSerializer.extend({
-  include: ['owner']
 });
